Checkpoint: moved logging_wrangler.py under core/
